+++ variables.tf
@@ -39,6 +39,54 @@ variable "alert_creation" {
   default     = "create_alerts_and_incidents"
   description = "(Optional) Must be one of two values. PagerDuty receives events from your monitoring systems and can then create incidents in different ways. Value `create_incidents` is default: events will create an incident that cannot be merged. Value `create_alerts_and_incidents` is the alternative: events will create an alert and then add it to a new incident, these incidents can be merged. This option is recommended."
 }
+variable "incident_urgency_rule" {
+  type = set(object(
+    {
+      during_support_hours = list(object(
+        {
+          type    = string
+          urgency = string
+        }
+      ))
+      outside_support_hours = list(object(
+        {
+          type    = string
+          urgency = string
+        }
+      ))
+      type    = string
+      urgency = string
+    }
+  ))
+  default = []
+}
+variable "support_hours" {
+  type = set(object(
+    {
+      type         = string
+      start_time   = string
+      end_time     = string
+      days_of_week = list(number)
+      time_zone    = string
+    }
+  ))
+  default = []
+}
+variable "scheduled_actions" {
+  type = set(object(
+    {
+      at = list(object(
+        {
+          name = string
+          type = string
+        }
+      ))
+      to_urgency = string
+      type       = string
+    }
+  ))
+  default = []
+}
 
 #################################
 # PagerDuty Service Integration